make sections into components
